fix!: return iterators from synchronous sources
Crossing async boundaries is not free.  If a synchronous iterator
is passed to a function, it should not force the caller to await on
the result.

This PR updates some of the basic `it-*` modules to not create unecessary
async when they don't have to.

The return types of these functions are now derived from the input types
so if you have linting rules that forbid awaiting on non-thenables those
awaits will need to be removed.

BREAKING CHANGE: if you pass a synchronous iterator to a function it will return a synchronous generator in response